# app/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.25.0

package db

import (
	"database/sql"
	"time"
)

type Advertisement struct {
	ID      int32     `json:"id"`
	Title   string    `json:"title"`
	StartAt time.Time `json:"start_at"`
	EndAt   time.Time `json:"end_at"`
}

type AdvertisementCond struct {
	ID              int32 `json:"id"`
	AdvertisementID int32 `json:"advertisement_id"`
	CondID          int32 `json:"cond_id"`
}

type Cond struct {
	ID       int32         `json:"id"`
	AgeStart sql.NullInt32 `json:"age_start"`
	AgeEnd   sql.NullInt32 `json:"age_end"`
}

type CondCountry struct {
	ID        int32 `json:"id"`
	CondID    int32 `json:"cond_id"`
	CountryID int32 `json:"country_id"`
}

type CondGender struct {
	ID       int32 `json:"id"`
	CondID   int32 `json:"cond_id"`
	GenderID int32 `json:"gender_id"`
}

type CondPlatform struct {
	ID         int32 `json:"id"`
	CondID     int32 `json:"cond_id"`
	PlatformID int32 `json:"platform_id"`
}

type Country struct {
	ID   int32  `json:"id"`
	Name string `json:"name"`
	Code string `json:"code"`
}

type Gender struct {
	ID   int32  `json:"id"`
	Name string `json:"name"`
	Code string `json:"code"`
}

type Platform struct {
	ID   int32  `json:"id"`
	Name string `json:"name"`
}