# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "請⽤ Golang 設計並且實作⼀個簡化的廣告投放服務，該服務應該有兩個 API，⼀個⽤於產⽣廣告，⼀個⽤於列出廣告。每個廣告都有它出現的條件(例如跟據使⽤者的年齡)，產⽣廣告的 API ⽤來產⽣與設定條件。投放廣告的 API 就要跟據條件列出符合使⽤條件的廣告",
        "title": "Dcard Backend Intern 2024",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/ad": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertisement"
                ],
                "summary": "列出符合可⽤和匹配⽬標條件的廣告",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "年齡條件",
                        "name": "age",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "M",
                            "F"
                        ],
                        "type": "string",
                        "description": "性別條件 (M/F)",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "國家條件 (參考 ISO_3166-1 alpha-2)",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "android",
                            "ios",
                            "web"
                        ],
                        "type": "string",
                        "description": "平台條件",
                        "name": "platform",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertisement"
                ],
                "summary": "產⽣廣告資源",
                "parameters": [
                    {
                        "description": "廣告內容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.Advertisement"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "handlers.Advertisement": {
            "type": "object",
            "required": [
                "endAt",
                "startAt",
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "x-order": "0",
                    "example": "AD 55"
                },
                "startAt": {
                    "type": "string",
                    "x-order": "1",
                    "example": "2023-12-10T03:00:00.000Z"
                },
                "endAt": {
                    "type": "string",
                    "x-order": "2",
                    "example": "2023-12-31T16:00:00.000Z"
                },
                "conditions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.AdvertisementCondition"
                    },
                    "x-order": "3"
                }
            }
        },
        "handlers.AdvertisementCondition": {
            "type": "object",
            "properties": {
                "ageStart": {
                    "type": "integer",
                    "x-order": "0",
                    "example": 20
                },
                "ageEnd": {
                    "type": "integer",
                    "x-order": "1",
                    "example": 30
                },
                "gender": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "x-order": "2",
                    "example": [
                        "M"
                    ]
                },
                "country": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "x-order": "3",
                    "example": [
                        "TW",
                        "JP"
                    ]
                },
                "platform": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "x-order": "4",
                    "example": [
                        "android",
                        "ios"
                    ]
                }
            }
        }
    }
}